+++ TODO.md
@@ -1,9 +1,9 @@
 # TODO
 
 * Smash Types.hs into
-  1. Environments
-  2. Terms = Expr + Pred + Const + etc. 
-  3. Sorts
-  4. Refinements
-  5. Constraints
-  6. Substitutions
+  + Sorts
+  - Environments
+  - Terms = Expr + Pred + Const + etc. 
+  - Substitutions
+  - Refinements
+  - Constraints
